requirements.txt
# Core FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Configuration and environment
python-dotenv==1.0.0
pydantic==2.5.0

# AI/LLM dependencies - Gemini Flash 2.0 Pro
google-generativeai==0.3.2

# Utilities
requests==2.32.4 